backend_flask/requirements.txt
flask
tensorflow
pillow
flask_cors
pyngrok